+++ packages/shared/src/index.ts
@@ -1,4 +1,5 @@
 export * from './logging/logger';
 export * from './logging/http-request';
+export * from './http/access-log.middleware';
 export * from './security/headers';
 export * from './types/domain';